requirements.txt
absl-py==1.0.0
contextlib2==21.6.0
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
ml-collections==0.1.0
numpy==1.21.4
Pillow==8.4.0
pyparsing==3.0.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
